[IT-1186] Add shared Nextflow development account under ScienceOU (#131)
I'm opening this PR to create a development AWS account for running Nextflow workflows. The intent is that the development account will used to determine the minimal policies, etc. required for the production account. This assumes that we will have more access to the development account than the production account. This account will also be used to test deployment from the `develop` branch in the [provisioner repository](https://github.com/Sage-Bionetworks/aws-nextflow-workflows). 

See #130 for the PR creating the corresponding development account.