[IT-1186] Add shared Nextflow production account under ScienceOU (#130)
I'm opening this PR to create a production AWS account for running Nextflow workflows. The intent is that the production account will be used on PHI data, so it needs to be HIPAA compliant. I expect there to be manual steps once this PR is merged. 

See #131 for the PR creating the corresponding development account.