improve ofn performance
The CI build job takes 20 mins to create a single AWS which seems really slow.
Attempt to make it faster by running more CI deployments in parallel.

The benefit of this is that it should make deployments faster, the
drawback is that it will make logs more difficult to read because logs
will be all jumpbled up.

The strategy is to run in parallel only on CI deployment, when manually
running deployments the concurrent executors can be set to 1 to get more
sane logs.