[IT-1186] Set up JumpCloud access for Nextflow Dev account (#136)
This PR builds on top of #131 and sets up access to the AWS account via JumpCloud. Since this is a development/sandbox account, it would be nice to be granted access to this role. This would allow me to create the role required for me to set up the provisioner repository, which needs to be configured to trust the service user mentioned in #135. 

That said, if you prefer setting up a `PowerUser` role, I can add a separate role in this PR so Sage IT can access the admin role and I (along with Tess) can access the `PowerUser` role.