[IT-1186] Set up JumpCloud access for Nextflow Prod account (#135)
This PR builds on top of #130 and sets up access to the AWS account via JumpCloud. I don't expect to be granted admin access on this account. I'm thinking that Sage IT can log into the account to create the service user and role required for me to set up the provisioner repository.