[SC-320] Product version per environment
This is a refactor to de-couple SC product versions from the environment
that it gets deployed to.  The goal is to allow developers to add
a product to one environment at a time.  For example when I'm developing
a new SC product I may want to merge a change that only deploys to the
development environment for testing.