[SC-251] Fix SSM document policy (#218)
The SC user gets access denied when attempting to start a session
with SSM dcouments.  The problem was due to the SC user policy.

Mixing SSM document with the policy to restrict users to instances
does not work due to the resource tag condition on the policy.
We need to split the policy up so that the condition does not
apply to the SSM documents.

This PR also allows users to start an command on an instance without
having to open a shell to it by starting the session with
AWS-StartNonInteractiveCommand document.