[SC-304] Deprecate product versions (#228)
AWS cloudformation does not support product deprecation so we've been
doing it manually however that's been an error prone process.  This PR
uses the cloudformation product provider custom resource[1] to
automatically deprecate all versions of a product except for the latest
version.  Users will only be allowed to provision with the latest version
of a product in production environments which prevents confusion and
users provisioning with old version that may have bugs in them.