[SC-304] force update version on every deploy (#230)
The lambda for UpdateProductVersions will not run if there aren't
any changes for the resource.  That's how cloudformation
idempotency works.

The problem is that when a new version of a product is added the
UpdateProductVersions lambda does not execute which means that
the previous version of the product will not get deprecated.
In this case we need to force a change to the resource so that
CFN will execute the lambda on every deployment.  That will make
sure that all previous versions of a product is deprecated.