[SC-345] Fix missing actions (#259)
This is a potential fix for missing actions buttons.
Evidence shows that missing actions are only hapenning with
general Linux EC2 products.  My theory is that the Sub in the
product description is causing a new version of the product to be created
on every deploy which would cause the association between actions
and product versions to become out of sync.