Add TowerViewer SSO role (#282)
We want Tower users to have some limited view access into the production account. This PR creates a role bound to an inline policy that is a subset of the permissions in the `ViewOnlyAccess` AWS-managed policy. 

If there's a preference for a customer managed policy over the current inline policy, I just need an example to follow. 

@tthyer: I'm assuming that we don't need to maintain separate JumpCloud user groups for the production and develop `TowerViewer` roles. Do you agree?