[SC-373] Fix tag options (#354)
Tag options were failing to deploy because internal and external tag
option configurations shared the same CFN template.  This caused
the template to deploy duplicate tag options.

To fix this issue we change the template to optionally create the
tag options.  If it's not specified then don't create them.  This
allows us to completely remove the CostCenter tags from BMGFKI and
STRIDES service catalog accounts.  This also allows us to remove
the cost center tags for extnernal users.