Fix Guardduty
Guardduty got into a broke state due to PR #380.

The problem was that we aded ItProdOU to the list of accounts
That OU contained securitycentral which should not be added as
a guardduty member because it's already the guardduty master.

This change uses `*` to adds all accounts and excludes securitycentral
which will essentially enable guardduty for all accounts going
forward.

Note: in order to get guardduty out of the bad state it had to be
completely reset.