[IT-1234] Fix VPN full tunnel option (#386)
We've noticed that when we disable VPN split tunnel there is no outbound
route to the internet when connected with the AWS VPN client.  When the split
tunnel option is enabled then the internet works fine.

I believe the problem is that when we enable full tunnel (disable split
tunnel) all outbound traffic goes thru the VPN and since the VPN is currently
only associated to private subnets there is no outbound routes to the internet.

To fix this issue we need to add a public subnet with a NAT to the VPN to allow
an outbound route to the internet.