Remove public subnet
We are playing a game of wack-a-mole when attempting to update
the public subnet for the VPN because AWS has a hard time
updating all the dependencies on the subnet configuration.

In an attempt to start with a clean slate we are going to
delete the public subnet and recreate it.