Disable SCP region restriction for testing
We are getting a message that says that admins are denied access to route53
resources.  We suspect that our SCP region restriction is causing this
problem.  We disable to test whether this is indeed the issue.

2022-03-28 20:02:59] - develop/agora R53RecordSet AWS::Route53::RecordSet CREATE_FAILED API:
route53:GetHostedZone User: arn:aws:sts::XXXXXX5864:assumed-role/AWSReservedSSO_Administrator_f5dc1fd84ac184d1/YYYY@sagebase.org
is not authorized to access this resource