[IT-1671] Prevent users from escalating privileges (#413)
The PowerUserAccess job function policy gives our SSO roles the
ability to assume other roles in an AWS account.  Since our JC
users are mapped to SSO roles this would allow our JC users to
potentially escalate their privileges by assuming a more
privileged role.

We want our IDP to be the source of truth for access to AWS.  This
setup is a security hole because it allows the role to circumvent
how we manage AWS access in JC.  To prevent this from happening we add
an explict deny to not allow our JC users to assume other roles.