Prevent parallel AWS deployments (#423)
We use github actions to deploy cloudformation to AWS.  This deployment
is currently setup to run in parallel when multiple PRs are merged at
the same time.  Infrastructure deployments to AWS do not work therefore
we want to turns off github action parallel deployments to AWS.

This separates the GH action to two workflows.  The deployment
workflow will only run if there is a merge to the master branch and
the validate workflow passes.