[IT-1234] Fix VPN resource names (#447)
Using an index for resource names is not good because it's not
gauranteed to be in the same order on the next deployment.  This
can cause AWS to do many replacements which may force users to
reconnect to the VPN.  We change to more stack naming convention
to make it consistent across deployments.