Grant Matthias Courville (GenUI) access to agora2-dev (#493)
Until we have a functional environment up and running, I'd like Matt to have as much access to agora2-dev as possible so that he can assist with debugging the migration. This could include doing any or all the things you listed as necessary until we can get this working.