[IT-1921] Update costcenter tags for some D&T AWS accounts (#488)
AWS accounts can be tagged with `CostCenter` so that IT can report all AWS expenses under the account to a particular code if a code exists per Bruce's comment [on slack](https://sagebionetworks.slack.com/archives/CEFQD0KU1/p1652735680132299).  This is an effort to add the CostCenter tags and have a discussion about the added tags.