[IT-1904] Encrypt cloudwatch trails
To encrypt Cloudtrail we need to create a KMS key with a custom
policy[1].  Since the dependency between the KMS key and cloudtrail
is so tight it's best to create the key with the cloudtrail resource
so we create a custom template with both key and trail.

Cloudtrail only allows encryption on the master trail because that's
the only one that's posting to a bucket.

[1] https://docs.aws.amazon.com/awscloudtrail/latest/userguide/default-kms-key-policy.html