Remove templates for cloudtrail and config for external accounts (#568)
While fixing issue IT-1904 I noticed that the templates we had
setup to accomodate external accounts does not work for a few
reasons.  First, the intention of these templates was to append
policies to the S3 bucket but instead it will overwrite the
existing policies on the cloudtrail and config buckets.  Second,
we don't even need it since the existing policy setup by the
templates referenced in `080-aws-config-inventory/_template.yaml`
and `060-cloudtrail/_templates.yaml` already accommodate external
accounts.