[IT-2089] Setup centralized multi-account patching
We have setup a maintenance.yaml[1] cloudformation template to execute automated
EC2 patching during maintenance windows.  The problem with this setup is that its
setup on a per account basis.  This make it pretty cumbersome to manage and maintain
across multiple AWS accounts.  The better approach is to setup a centralized patching
center[2] that will go out and patch every instance in every AWS account.

[1] https://github.com/Sage-Bionetworks/aws-infra/blob/master/templates/maintenance.yaml
[2] https://aws.amazon.com/blogs/mt/scheduling-centralized-multi-account-multi-region-patching-aws-systems-manager-automation/