Create a system manager account (#577)
This PR is related to IT-1729 and IT-2089

We want to setup a few new system management services:
  * AWS instance scheduler service[1]
  * AWS centralized patching service[2]

These services will manage instances accross multiple AWS accounts
so we need a central account to deploy these apps to.  I propose we
create a new account called `org-sagebase-sysmancentral` for these
types of services.  That way we can scope their access to systems
operators.

[1] https://docs.aws.amazon.com/solutions/latest/instance-scheduler-on-aws/welcome.html
[2] https://aws.amazon.com/blogs/mt/scheduling-centralized-multi-account-multi-region-patching-aws-systems-manager-automation/