[IT-1967] Create an identity central account (#599)
The idea behind identity-central is to have an AWS account that’s basically
a shell account containing only IAM users and nothing else.  The purpose is
to allow these IAM users to access resources in other AWS accounts but not
resources in the identity-central account. Identity-central would be integrated
with our Jumpcloud IDP so that we can manage AWS IAM users via Jumpcloud.