[IT-2089] narrow resource group for patching (#607)
We apply tags to EC2 to identify which EC2s need patching.  The problem
is that the tags are also propogated to the EC2s attached volumes.
This means that the patch group query will return EBS volumes for
patching which will fail.  We narrow the scope of the query so that
only instances are returned to the patch manager to run patch update on.