[DCA-34] Allow deployment from multiple branches
For deployments where there is a more involved change to the infrastructure
we may want to manage the deployment of that change to one
account/environment (dev) before propogating it to other accounts (prod).

For this we setup multiple git branches (dev/stage/prod) to allow us to
propogate the deployments which means we need to allow OIDC to deploy
from multiple branches.