Setup viewer access to org-sagebase-ipam account (#752)
AWS recommends assigning non-overlapping CIDRs to VPCs[1].
We have been using the infra inventory wiki page[2] to keep track of
our network CIDRs across our entire organization.  That wiki page is
not really working anymore so we are going to replace it with info
from our org-sagebase-ipam account.  The info in ipam is dynamic
and will allow users to find unused CIDRs to assign to their
VPCs.

[1] https://docs.aws.amazon.com/vpc/latest/userguide/configure-your-vpc.html#vpc-cidr-blocks
[2] https://sagebionetworks.jira.com/wiki/spaces/IT/pages/352976898/Infrastructure+Inventory