[IT-2774] Update Service Catalog CostCenter tag list
Switch from getting valid tags values from static files to a rest
endpoint that provides updated tag values from the source of truth.

We use the sceptre-request-resolver[1] to get the cost center tags
as a string then we use the sceptre-json-resolver[2] to convert it
to a list object so that we can iterate on each item to setup tags
for the service catalog.

[1] https://github.com/Sceptre/sceptre-request-resolver
[2] https://github.com/Sceptre/sceptre-json-resolver

depends on Sceptre/sceptre-request-resolver#1 Sceptre/sceptre-json-resolver#1