[IT-2768] make OIDC template support multi repos
Our current GH OIDC template only sets up GH OIDC for a single
repository in an GH org.  This causes lots of duplication in the
orgf _task.yaml file.  This update will allow setting up OIDC for
multiple repos and branches which will allow us to reduce the number of
entries in the _task.yaml file.