[IT-2768] make OIDC template support multi repos (#859)
Our current GH OIDC template only sets up GH OIDC for a single
repository in an GH org.  This causes lots of duplication in the
orgf _task.yaml file.  This update will allow setting up OIDC for
multiple repos and branches which will allow us to reduce the number of
entries in the _task.yaml file.

We need to provide a provider role name to allow us to create dependent
PRs to repos that require the name to assume.  This will allow us to update with
minimum downtime.