Fix github OIDC configuration
There was a bug in PR #859 where we were setting '*' condition for
branches incorrectly.  According to the AWS docs[1], granting access
to all branches does not include the refs.

"The following example condition limits access to the defined GitHub
organization and repository, but grants access to any branch within
the repository"
```
"Condition": {
  "StringLike": {
    "token.actions.githubusercontent.com:aud": "sts.amazonaws.com",
    "token.actions.githubusercontent.com:sub": "repo:GitHubOrg/GitHubRepo:*"
  }
}
```

[1] https://docs.aws.amazon.com/IAM/latest/UserGuide/id_roles_create_for-idp_oidc.html