[IT-1532] Send monthly cloud-spend emails
Canary deployment of a lambda to send monthly email notifications
summarizing cloud spend and linking to confluence documentation on
how to access AWS Cost Explorer.

Restrict the recipient list to Sage IT as a canary test and to
review the content and formatting of the email.

Depends on Sage-Bionetworks-IT/lambda-finops-email-totals#5