Remove hard dependency from essentials stack (#984)
The older version of the essentials.yaml template contained
an `!ImportValue` reference which sets up a hard dependency
which makes it difficult to update values.  We update
to a newer version which takes in a value to allow users
to more easily change values thus making easier to do updates.

This also removes a VPC peering role which we no longer use since
we switched from VPC peering to the transit gateway.