[IT-3356] Remove Guardduty automation (#1099)
The guardduty automation was to enable guardduty and member
accounts thru the invitation system.  We have changed it so that
guardduty is now managed from our Organizations account[1]. 
This is a requirement to allow us to enable some of the Guardduty
features such as malware protection. There is no automation
support for the Organizations approach therefore Guardduty is
no longer managed with IaC.

[1] https://docs.aws.amazon.com/guardduty/latest/ug/guardduty_settingup.html#guardduty_enable-gd