[IT-2581] Refactor cloudtrail setup
Previously to enable cloudtrail to send data to a central AWS account
(logcentral) we need to setup cloudtrail in every member account to
enable this configuration.

AWS has now introduced a new feature that allows enabling `organizations
cloudtrail` without having to setup cloudtrail in every member account.
Instead it just need to be enabled from an organizations account or from
a deleaged admin account.

We are changing the cloudtrail setup to the latter option because it's
much simpler.  The only caveat is that we need to first manually setup
logcentral as an cloudtrail delegated admin account[1]

[1] https://docs.aws.amazon.com/awscloudtrail/latest/userguide/cloudtrail-add-delegated-administrator.html