[IT-2581] Remove Cloudtrail automation (#1130)
This reverts commit 5a2dfaacb3 and completely removes the cloudformation
automation for cloudtrail setup.  I have tried to setup cloudtrail's
`Organizations trail` using cloudformation however it doesn't seem
to be possible.

The problem is that enabling an organizations trail in a delegated
account (logcentral) also requires setting up cloudtrail in the
organizations account.  The cloudtrail in both accounts have circular
dependencies.  The two way dependency requires a coordinated
deployment where things need to be deployed in logcentral then in
organizations then again in logcentral.  This is does not seem possible
with our cloudformation setup.

Setting up from the cloudtrail AWS console is possible because I think
AWS is doing something special, possibily using a step function, to
coordinate the deployment between the two accounts.
We are removing this attempt at setting up  cloudtrail organizations
trail and will be setting it up manually from the AWS console.