[IT-3554] Add additional Service Catalog permissions (second attempt)
Updating provisioned products with changes to tags is failing with permissions
errors. RestrictBucketDownloadRegionFunction updates fail on 'lambda:ListTags',
and the associated role fails on'iam:UntagResource'. Allow SC roles to list tags
on any lambda, and to update tags on their own provisioned resources.

Create a separate role in scipool-dev for manually testing permission changes
before applying the changes to scipool-prod.

The 'aws:ResourceTag/foo' condition key lets us compare against the value
of the 'foo' tag on the target resource. Here we are comparing the
'provisioningPrincipalArn' SC service tag value against the 'PrincipalArn' from
the request.

https://docs.aws.amazon.com/IAM/latest/UserGuide/access_tags.html#access_tags_control-resources
https://docs.aws.amazon.com/IAM/latest/UserGuide/reference_policies_condition-keys.html#condition-keys-resourcetag
https://docs.aws.amazon.com/IAM/latest/UserGuide/reference_policies_condition-keys.html#condition-keys-principalarn