[IT-3697] Test the latest docker AMI before tagging a new version (#1170)
Test the latest AMI created by packer-amazonlinux-docker before tagging a new version by referencing the master AMI in the product template on the master branch of service-catalog-library and referencing that product template in the dev account here.

If the test works then packer-amazonlinux-docker will be tagged with a new version, the tagged AMI will be referenced in the product template and service-catalog-library will be tagged with a new version. Finally, new products will be created here from the newly tagged product template.