[IT-3619] Enable AWS config service (#1195)
This is a partial revert of 9f5bb53

We need to enabled AWS config service in every organization member account
to collect config data.  This will allow the aggregator account (logcentral)
to collect and aggregate the config data from member accounts, visualize
that data and store it in a central account.