[IT-3758] Setup a bucket for four points
Four point's use cases are broader than initialally anticipated when
setting up a storage location for them in PRs #1189 and #1197

This unwinds changes in those PRs.  Instead of giving 4points access
to a folder in an existing bucket we create a whole new bucket for
them and give them reqd and write access to the new bucket.