Fix public read bucket
This is a fix for PR #1202 which I tried to fix in PR #1225 but now
encountered the following error..

```
develop/sc-product-ec2-linux-docker-notebook sc-product-ec2-linux-docker-notebook
AWS::CloudFormation::Stack UPDATE_IN_PROGRESS Transform AWS::Include failed with:
S3 URI must reference a valid S3 object to which you have access.
```

Thich i think means that the bucket is still not setup as a public
read bucket.

This fix is derived from the example in AWS docs to setup an S3 public bucket[1]

[1] https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-s3-bucket.html#aws-resource-s3-bucket--examples--Granting_public_access_to_S3_buckets