[IT-4244] update SSO ImageCentralAmiLibrarian policy
We switch from using packer to using the AWS image builder to build
AMIs (in issue IT-3831).  This change requires AMI librarians to have
a different set of access to build and deploy AMIs with cloudformation
templates.

Note: This new set of policies was auto generated using the cfnlp[1]
app.

[1] https://github.com/iann0036/aws-leastprivilege