bugfix: ignore notices from the CDK CLI
The deploy pipeline is failing due to jq being unable to parse the
output from the CDK CLI. Generating the template locally creates a
valid JSON file with 1091 lines and prints a notice to stderr. In the
deploy pipeline jq failes to parse line 1093. It seems that in the
github-runner environmunt stderr is redirected to stdout, causing
the notice to be written to the end of the file.

Disable printing CDK CLI notices in the deploy pipeline.