[IT-4190] Suppress security hub rule (#1343)
I recommend we suppress this finding:
  5.1 Ensure no Network ACLs allow ingress from 0.0.0.0/0 to remote server administration ports

It’s typically normal to allow NACL ingress from all ports.
This is in fact how the default AWS VPC is setup.

I think it’s already cumbersome to configure access in AWS, closing off
NACL ports without an explicit reason would just makes it more
cumbersome and confusing to configure access to resources.  There is an
additional mechanism, security groups, that would need to be setup to
allow access to resources.