[IT-3831] Update SSO read only access (#1383)
This reverts commit 58070be5e7dc39cd2fc88903409d4e1113a55536.

Try this again after making a request to increase IAM quota from the
default 10 managed policies per role to 20 per role.

Note: quota increase requests cannot be made at the org level, it needs
to be done in each member account and there is no automation for making
the request.  We need to manually login to every account to make the request