[IT-4347] Remove deployment of CloudWatch-CrossAccountSharingRole (#1406)
CloudWatch-CrossAccountSharingRole resource is being deployed multiple times
which is causing failures during creation of new AWS accounts.

I believe what is happening is that the AWS cloudwatch service is creating
a CloudWatch-CrossAccountSharingRole when we vend new AWS accounts thru
our CI/CD.  Then org-formation is deploying the duplicate role later in
the vending process which causes the failure.

Since AWS is already creating the required role I guess the fix would
be to remove the deployment of that role from our org-formation setup.