Throttle Sceptre deployment (#1498)
AWS is throttling cloudformation deployments to service catalog
accounts because there are too many requests to the service catalog
APIs. This is causing deployments to fail and put service catalog in
a bad state.

This reverts the temporary disabling of service catalog deployments
in PR #1496.  It also installs a development version of Sceptre containing
a max-concurrency option that will throttle sceptre requests to avoid
the AWS throttling.