[IT-4720] revert to sceptre parallel deploy
We changed Sceptre to deploy service catalog cloudformation templates
serially in PR #1498 to get around an AWS throttling issue.

AWS has confirmed that a change on their end caused this problem
and they have reverted their change.  From that we are reverting
our change so that we can deploy cloudformation in parallel again.